Jedis 대기열+활성화큐 구현. QueueManager(WaitingQueue, ActiveQueue) 구분. WaitingQueue와 ActiveQueue는 추상화
