+++ src/main/java/hhplus/ticketing/domain/token/models/TokenStatus.java
@@ -1,4 +1,4 @@
-package hhplus.ticketing.domain.watingqueue.models;
+package hhplus.ticketing.domain.token.models;
 
 public enum TokenStatus {
     ACTIVE, WAITING, EXPIRED;
