--- src/test/java/hhplus/ticketing/watingqueue/integration/WaitingQueueJedisTest.java
@@ -1,93 +1,0 @@
-package hhplus.ticketing.watingqueue.integration;
-
-import hhplus.ticketing.base.exceptions.InvalidTokenException;
-import hhplus.ticketing.domain.watingqueue.components.QueueManager;
-import hhplus.ticketing.domain.watingqueue.components.WaitingQueueService;
-import hhplus.ticketing.domain.watingqueue.models.Token;
-import hhplus.ticketing.domain.watingqueue.models.TokenStatus;
-import hhplus.ticketing.domain.watingqueue.models.WaitingInfo;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import java.time.LocalDateTime;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-@SpringBootTest
-public class WaitingQueueJedisTest {
-
-    @Autowired
-    QueueManager queueManager;
-    @Autowired
-    WaitingQueueService waitingQueueService;
-    @Test
-    @DisplayName("없는 토큰을 조회할 경우 예외 처리")
-    void query_invalid_token(){
-        Token invalidToken = new Token(1, "INVALID-TOKEN", 1, TokenStatus.WAITING, LocalDateTime.now());
-
-
-        assertThrows(InvalidTokenException.class, () ->
-                waitingQueueService.query(invalidToken));
-
-    }
-
-    @Test
-    @DisplayName("유저ID로 대기열에 등록하면 토큰을 받는다.")
-    void receive_token_when_register_in_waiting_queue_with_userID(){
-        long userId = 1;
-
-        Token token = waitingQueueService.register(1, userId, LocalDateTime.now());
-        WaitingInfo waitingInfo = waitingQueueService.query(token);
-
-        assertThat(waitingInfo.userId()).isEqualTo(userId);
-    }
-
-    @Test
-    @DisplayName("토큰으로 대기열을 조회하면, 현재 대기순서를 알 수 있다.")
-    void query_waiting_number_by_token(){
-        long userId1 = 100;
-        long userId2 = 200;
-
-        Token token1 = waitingQueueService.register(1, userId1, LocalDateTime.now());
-        Token trgToken = waitingQueueService.register(1, userId2, LocalDateTime.now());
-
-        WaitingInfo info = waitingQueueService.query(trgToken);
-
-        assertThat(info.waitingNo()).isEqualTo(2);
-    }
-
-    @Test
-    @DisplayName("입장 순서가 되면 토큰을 활성화시킨다.")
-    void activate_token_when_waiting_is_over(){
-
-        long userId = 1;
-        Token token = waitingQueueService.register(1, userId, LocalDateTime.now());
-
-        assertThat(queueManager.checkActive(token)).isFalse();
-
-        queueManager.activate(token, LocalDateTime.now());
-
-        assertThat(queueManager.checkActive(token)).isTrue();
-    }
-
-    @Test
-    @DisplayName("토큰 유효시간이 지나면 만료시킨다.")
-    void deactivate_token_when_time_up(){
-
-        int n = 1;
-        long concertId = 1L;
-        long userId = 1L;
-        LocalDateTime issuedAt = LocalDateTime.now();
-        LocalDateTime timeAlreadyExpired = issuedAt.minusMinutes(10);
-
-        Token token = waitingQueueService.register(concertId, userId, issuedAt);
-        queueManager.activateTokensByTimeOrder(concertId, n);
-        assertThat(queueManager.checkActive(token)).isTrue();
-
-        queueManager.deactivate(token);
-        assertThat(queueManager.exists("ACTIVATE:CONCERT_ID" + token.getConcertId(), token)).isFalse();
-    }
-}
